move: Data와Domain 관련 Repository 각 폴더로 분리
