+++ DevFriends/DevFriends/Domain/Repositories/UserRepository.swift
@@ -10,18 +10,3 @@ import Foundation
 protocol UserRepository: ContainsFirestore {
     func fetchUser() async throws -> User
 }
-
-final class DefaultUserRepository {}
-
-extension DefaultUserRepository: UserRepository {
-    func fetchUser() async throws -> User {
-        // MARK: user를 나중에 어떻게 가져올지 논의해보기
-        guard let uid = UserDefaults.standard.object(forKey: "uid") as? String
-        else { fatalError("UID was not stored!!") }
-        let userSnapshot = try await firestore.collection("User").document(uid).getDocument()
-        
-        let user = try userSnapshot.data(as: User.self)
-        
-        return user
-    }
-}
